Fonts now scaling with floating point precision

+++ Core/UI/Inc/steering_types.h
@@ -56,7 +56,7 @@ typedef struct {
     uint32_t swoc_elem_lb_color;
     uint32_t swoc_elem_bg_color;
     const Olivec_Font *swoc_elem_font;
-    const size_t swoc_elem_font_size;
+    const float swoc_elem_font_size;
     const Olivec_Boundaries swoc_elem_boundaries;
 } UI_elem_t;
